refactor: 정신없는 파일 구조를 개선했습니다.
headerAnd~~ 묶어서 관리했던 것을 다 분리해서 components/header 폴더를 최적화했습니다.
