--- frontend/src/components/header/HeaderAndTap.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import Header from './Header';
-import HeaderTap from './HeaderTap';
-
-export default function HeaderAndTap(props) {
-  return (
-    <>
-      <Header />
-      <HeaderTap
-        ancestorMenuTree={props.ancestorMenuTree}
-        currentTabContents={props.currentTabContents}
-        handleBoardChange={props.handleBoardChange}
-      />
-    </>
-  );
-}
