+++ frontend/src/components/header/Navigation.js
@@ -4,20 +4,16 @@ import styled from 'styled-components';
 import theme from '../../styles/Theme';
 
 // ancestorMenuTree 상위 메뉴를 받아와 링크로 만들어준다.
-export default function HeaderTap(props) {
-  const changeTab = index => {
-    props.handleBoardChange(props.currentTabContents[index].name);
-  };
-
+export default function Navigation(props) {
   return (
-    <HeaderTabContainer>
+    <NavigationContainer>
       <AncestorMenu>
         {props.ancestorMenuTree.map((menu, idx) => (
           <>
             <AncestorMenuLink to={menu.link} key={`ancestor${idx}`}>
               {menu.name}
             </AncestorMenuLink>
-            <RightAngleBracket>{' > '}</RightAngleBracket>
+            <RightAngleBracket key={`bracket${idx}`}>{' > '}</RightAngleBracket>
           </>
         ))}
       </AncestorMenu>
@@ -25,23 +21,25 @@ export default function HeaderTap(props) {
         {props.currentTabContents.map((menu, idx) => (
           <>
             <CurrentMenuTabContents
-              onClick={() => changeTab(idx)}
+              to={menu.link}
               key={`current${idx}`}
               accent={menu.accent}
             >
               {menu.name}
             </CurrentMenuTabContents>
+            <VeticalBar key={`veticalBar${idx}`}>
+              {idx < props.currentTabContents.length - 1 ? ` | ` : ''}
+            </VeticalBar>
           </>
         ))}
       </CurrentMenuTab>
-    </HeaderTabContainer>
+    </NavigationContainer>
   );
 }
 
-const HeaderTabContainer = styled.div`
+const NavigationContainer = styled.div`
   width: 100%;
   height: 160px;
-  margin-bottom: 60px;
 `;
 
 const AncestorMenu = styled.nav`
@@ -50,7 +48,7 @@ const AncestorMenu = styled.nav`
   font-family: 'GmarketSansMedium', sans-serif;
   font-style: normal;
   font-weight: 500;
-  font-size: 20px;
+  font-size: ${theme.fontSizes.paragraph};
   line-height: 120%;
   text-decoration: none;
 `;
@@ -70,26 +68,21 @@ const CurrentMenuTab = styled.nav`
   text-align: center;
 `;
 
-const CurrentMenuTabContents = styled.span`
-  margin-left: 10px;
-  padding-left: 10px;
-  border-left: 1px solid ${theme.colors.white};
-
+const CurrentMenuTabContents = styled(Link)`
   color: ${theme.colors.white};
   font-family: ${props => (props.accent ? `GmarketSansMedium` : 'Pretendard')},
     sans-serif;
   font-style: normal;
   font-weight: ${props => (props.accent ? 500 : 300)};
-  font-size: 30px;
+  font-size: ${theme.fontSizes.title};
   line-height: 120%;
   text-decoration: none;
-  &:hover {
-    cursor: pointer;
-  }
+`;
 
-  &:first-child {
-    margin-left: 0;
-    padding-left: 0;
-    border-left: none;
-  }
+const VeticalBar = styled.span`
+  color: ${theme.colors.white};
+  font-family: sans-serif;
+  font-weight: 100;
+  font-size: ${theme.fontSizes.title};
+  text-decoration: none;
 `;
