--- frontend/src/components/header/TabContentByIndex.js
@@ -1,11 +1,0 @@
-import CabinetRentWindow from './../manage/CabinetRentWindow';
-import MemberAuthorizeWindow from './../manage/MemberAuthorizeWindow';
-import MemberInfoWindow from './../manage/MemberInfoWindow';
-import UmbrellaRentWindow from './../manage/UmbrellaRentWindow';
-
-export const TabContentByIndex = {
-  0: <MemberAuthorizeWindow />,
-  1: <MemberInfoWindow />,
-  2: <UmbrellaRentWindow />,
-  3: <CabinetRentWindow />,
-};
