+++ frontend/src/constants/ManageMenu.js
@@ -1,0 +1,11 @@
+import CabinetRentWindow from '../components/manage/CabinetRentWindow';
+import MemberAuthorizeWindow from '../components/manage/MemberAuthorizeWindow';
+import MemberInfoWindow from '../components/manage/MemberInfoWindow';
+import UmbrellaRentWindow from '../components/manage/UmbrellaRentWindow';
+
+export const ManageMenu = {
+  0: <MemberAuthorizeWindow />,
+  1: <MemberInfoWindow />,
+  2: <UmbrellaRentWindow />,
+  3: <CabinetRentWindow />,
+};
