+++ frontend/src/pages/Manage.js
@@ -1,30 +1,32 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import HeaderAndManageTap from '../components/header/HeaderAndManageTap';
-import { TabContentByIndex } from '../components/header/TabContentByIndex';
+import { ManageMenu } from '../constants/ManageMenu';
 import theme from '../styles/Theme';
+import Header from '../components/header/Header';
+import ManageTab from '../components/header/ManageTab';
 
 function Manage(props) {
-  const [tabContentIndex, setTabContentIndex] = useState(0);
+  const [memu, setMemu] = useState(0);
 
   const changeTabContent = index => {
-    setTabContentIndex(index);
+    setMemu(index);
   };
 
   const currentTabContents = [
-    { name: '회원 승인', accent: tabContentIndex === 0 },
-    { name: '회원 정보', accent: tabContentIndex === 1 },
-    { name: '우산 대여', accent: tabContentIndex === 2 },
-    { name: '사물함 대여', accent: tabContentIndex === 3 },
+    { name: '회원 승인', accent: memu === 0 },
+    { name: '회원 정보', accent: memu === 1 },
+    { name: '우산 대여', accent: memu === 2 },
+    { name: '사물함 대여', accent: memu === 3 },
   ];
 
   return (
     <ManageContainer>
-      <HeaderAndManageTap
+      <Header />
+      <ManageTab
         currentTabContents={currentTabContents}
         changeTabContent={changeTabContent}
       />
-      {TabContentByIndex[tabContentIndex]}
+      {ManageMenu[memu]}
     </ManageContainer>
   );
 }
